feat: sync market colors once on market list
