+++ src/helpers/buildMarketColors.ts
