+++ src/helpers/hasMarketColors.ts
