+++ src/hooks/index.ts
@@ -25,8 +25,6 @@ export { default as useExpandableOutcomes } from './useExpandableOutcomes';
 
 export { default as useFantasyTokenTicker } from './useFantasyTokenTicker';
 
-export { default as useMarketColors } from './useMarketColors';
-
 export { default as useMarkets } from './useMarkets';
 
 export { default as useMarketForms } from './useMarketForms';
